recipes-importer/requirements.txt
pandas
openpyxl
mysql-connector-python